minor changes lib and withdraw + subscribe only subscribes
