+++ subscriptions/client/src/instructions/create_subscription.rs
@@ -1,4 +1,5 @@
 use {
+    crate::*,
     anchor_lang::{prelude::Pubkey, solana_program::sysvar, InstructionData},
     anchor_spl::{associated_token, token},
     solana_client_helpers::{Client, ClientResult, RpcClient, SplToken},
@@ -12,8 +13,6 @@ use {
     },
 };
 
-//  pub rent: Sysvar<'info, Rent>,
-
 fn create_subscription(
     client: &Client,
     subscription_bank: Pubkey,
@@ -25,7 +24,7 @@ fn create_subscription(
     is_active: bool,
     subscription_id: String,
 ) -> ClientResult<()> {
-    let create_payment_ix = Instruction {
+    let create_subscription_ix = Instruction {
         program_id: subscriptions_program::ID,
         accounts: vec![
             AccountMeta::new(client.payer_pubkey(), true),
@@ -49,35 +48,16 @@ fn create_subscription(
         .data(),
     };
 
-    // let top_up_payment_ix = Instruction {
-    //     program_id: payments_program::ID,
-    //     accounts: vec![
-    //         AccountMeta::new_readonly(associated_token::ID, false),
-    //         AccountMeta::new(escrow, false),
-    //         AccountMeta::new(payment, false),
-    //         AccountMeta::new_readonly(mint, false),
-    //         AccountMeta::new_readonly(recipient, false),
-    //         AccountMeta::new_readonly(sysvar::rent::ID, false),
-    //         AccountMeta::new(client.payer_pubkey(), true),
-    //         AccountMeta::new(sender_token_account, false),
-    //         AccountMeta::new_readonly(system_program::ID, false),
-    //         AccountMeta::new_readonly(token::ID, false),
-    //     ],
-    //     data: payments_program::instruction::TopUpPayment {
-    //         amount: LAMPORTS_PER_SOL,
-    //     }
-    //     .data(),
-    // };
+    send_and_confirm_tx(
+        client,
+        &[create_subscription_ix],
+        "create_payment_with_top_up".to_string(),
+    )?;
 
-    // send_and_confirm_tx(
-    //     client,
-    //     &[create_payment_ix, top_up_payment_ix],
-    //     "create_payment_with_top_up".to_string(),
-    // )?;
-    //
-    //     println!(
-    //         "queue: https://explorer.solana.com/address/{}?cluster=custom",
-    //         payment_queue
-    //     );
+    println!(
+        "queue: https://explorer.solana.com/address/{}?cluster=custom",
+        subscription_queue
+    );
+
     Ok(())
 }
