+++ subscriptions/programs/subscriptions/src/contexts/disburse_payment.rs
@@ -1,7 +1,10 @@
 use {
     crate::state::*,
     anchor_lang::prelude::*,
-    clockwork_sdk::{queue_program::accounts::Queue, CrankResponse},
+    clockwork_sdk::{
+        queue_program::{self, accounts::Queue, QueueProgram},
+        CrankResponse,
+    },
 };
 
 #[derive(Accounts)]
@@ -19,13 +22,18 @@ pub struct DisbursePayment<'info> {
     pub subscription: Account<'info, Subscription>,
     #[account(address = Queue::pubkey(subscription.key(), "subscription".into()))]
     pub subscriptions_queue: Account<'info, Queue>,
+
+    #[account(address = queue_program::ID)]
+    pub clockwork_program: Program<'info, QueueProgram>,
 }
 
 impl<'info> DisbursePayment<'_> {
-    pub fn process(&mut self) -> Result<CrankResponse> {
+    pub fn process(&mut self, bump: u8) -> Result<CrankResponse> {
         let Self {
             subscriber,
             subscription,
+            subscriptions_queue,
+            clockwork_program,
             ..
         } = self;
 
@@ -40,6 +48,19 @@ impl<'info> DisbursePayment<'_> {
             }
             None => {
                 subscriber.is_subscribed = false;
+                clockwork_sdk::queue_program::cpi::queue_stop(CpiContext::new_with_signer(
+                    clockwork_program.to_account_info(),
+                    clockwork_sdk::queue_program::cpi::accounts::QueueStop {
+                        authority: subscription.to_account_info(),
+                        queue: subscriptions_queue.to_account_info(),
+                    },
+                    &[&[
+                        SEED_SUBSCRIPTION,
+                        subscription.owner.as_ref(),
+                        subscription.subscription_id.as_bytes(),
+                        &[bump],
+                    ]],
+                ))?;
             }
         }
