+++ subscriptions/programs/subscriptions/src/contexts/mod.rs
@@ -1,9 +1,13 @@
+pub mod create_subscriber;
 pub mod create_subscription;
+pub mod deposit;
 pub mod disburse_payment;
 pub mod subscribe;
-pub mod unsubscribe;
+pub mod withdraw;
 
+pub use create_subscriber::*;
 pub use create_subscription::*;
+pub use deposit::*;
 pub use disburse_payment::*;
 pub use subscribe::*;
-pub use unsubscribe::*;
+pub use withdraw::*;
